+++ src/org/spacebar/escape/common/Player.java
@@ -1,7 +1,7 @@
 package org.spacebar.escape.common;
 
 public class Player extends Entity {
-    public Player(int x, int y, int d) {
+    public Player(int x, int y, byte d) {
         super(x, y, d);
 
         iAmPlayer();
